requirements.txt
requests
celery[redis]
sqlalchemy
pandas
geopandas
mkdocs-material
python-decouple
